Fix multithreaded visibility issue with TransformExecutor
Use an AtomicReference to ensure the thread the Executor is being run on
will be visible to other threads.

Add a checkState to ensure that the executor will not run more than
once.

Stop setting the active thread to null, as the Transform Executor should
be discarded from any references and not hold the overall execution.