Improve Callback scheduling in InProcessSideInputContainer
This removes the logic required to fill a SideInput with empty contents
from the read and moves it into the Cache Loader. This stops the
container for setting an emptiness callback every time a PCollectionView
is requested. Doing so also requires the write logic to provide a complete
SettableFuture that contains all of the values already set, as otherwise
once after the time output must have been produced, all updates would
instead return an empty value.