Add withElements to CommittedBundle
The unprocessed elements of a partially completed bundle must be
placed in a new bundle to be processed at a later time. The bundle in
which they are processed should also have identical properties to the
bundle which the elements were initially present in.

withElements provides a simple way to create a "copy" of a bundle that
contains different elements.