Use an AtomicReference in InProcessSideInputContainer
This ensures that the timing associated with setting an side input to
being empty can only occur if the contents of the Side Input have never
been set. It ensures that no caller ever sees an intermediate state of
the side input.

Throw an exception while trying to retrieve the contents of a SideInput
that is not ready.