Use custom Assertions in Spark Streaming Code
Spark Streaming Side Inputs behave differently than the Beam Model. As
the underlying implementation of PAssert is based on side inputs, this
means that Streaming Spark Tests that use SideInputs as the actuals are
non-portable.

Modify FlattenStreamingTest, KafkaStreamingTest, and
SimpleStreamingWordCountTest to group all of the contents of the
asserted PCollection into a single key, and assert the contents of that
concatenation, rather than doing so via PAssert and Side Input.