Document the thread hostility of DoFn output methods
For a single DoFn instance, contexts provided to that instance may share
objects that cannot be interacted with in a concurrency-safe manner.
This requires users to ensure that their code interacts with output
synchronously, and that calls to output cannot be made after a
ProcessElement, StartBundle, or FinishBundle method completes.