Add an Example Pipeline Demonstrating Unsafety
When a pane is output in Accumulating Mode, the _next_ pane will be
output with an output time independent of any elements within that pane.
This can lead to confusing behavior where elements cannot be reassigned
to (say) their original timestamp.

This is even more problematic if elements are rewindowed downstream of a
GroupByKey, as the window the elements appear to be in will change based
on the upstream triggering.