Inject number of Shards as a Side Input to Write
This permits users to pass a PTransform from the input elements to the
number of shards instead of requiring a constant amount. This enables
sharding to be determined based on the input data rather than a constant
value.

Tests still should be written. This can also convert the DirectRunner to
inject a custom sharding strategy in its override, rather than relying
on custom bundling.