Condense POutput Finalization methods
Provide the name of the transform that originally output the PValue or
is currently being finalized to finalization methods.

Remove getProducingTransformInternal, recordAsOutput from PValueBase.

Update PValueBase.recordAsOutput() to explicitly set the default name if
unset. Update PValue.finishSpecifying to set the
name to the default, and move to an empty implementation of
finishSpecifyingOutput.

PValues will not have a name set until they are either output as part of
a composite POutput or consumed as an input.