Genericise TransformEvaluatorFactories
Concrete factories need to know the type of executable they are
evaluating. However, the registry and the majority of the transform
executors don't need to have access to the specifics, only the
information required to determine which evaluator to use. The executable
can be passed through to the EvaluatorFactory without inspecting it.